Dataset generation
I made the first iteration of the dataset based on the boston bluebikes data from 2105 onwards, the weather data from NOAA and bostons GIS data. We still need to work on adding the business zoning aspect to the dataset. I have a basic outline for that but I haven't finished that part yet